Updated the example to correctly calculate the camera bounds.
Setting up the camera for the maps has been split into two methods, one for isometric maps and another for orthographic.
The isometric method can now correctly set the bounds, taking into account the first tile placement and determines the width
of the map either side of the x axis "0".